refactor: rearrange the folders
