--- src/services/__init__.py
@@ -1,3 +1,0 @@
-import os 
-import sys
-sys.path.append(os.path.dirname(__file__))
